fix : save load

+++ Return of the Blossoming Blade/Assets/Script/CutScene/Blossoming/CutScene1.cs
@@ -41,8 +41,8 @@ public class CutScene1 : MonoBehaviour
             PlayerPrefs.SetFloat("playerHP", PlayerPrefs.GetFloat("save"+ saveNum + "PlayerHP"));
             PlayerPrefs.SetFloat("playerMP", PlayerPrefs.GetFloat("save"+ saveNum+"PlayerMP"));
             //øµæ‡
-            PlayerPrefs.SetFloat("havePosion", PlayerPrefs.GetFloat("save"+ saveNum + "HavePosion"));
-            PlayerPrefs.SetFloat("maxPosion", PlayerPrefs.GetFloat("save"+ saveNum + "MaxPosion"));
+            PlayerPrefs.SetInt("havePosion", PlayerPrefs.GetInt("save"+ saveNum + "HavePosion"));
+            PlayerPrefs.SetInt("maxPosion", PlayerPrefs.GetInt("save"+ saveNum + "MaxPosion"));
 
             //∏  ¿Ãµø
             TransferMap[] temp2 = FindObjectsOfType<TransferMap>();
